Break out init routine to parent class
